+++ src/app/pages/custom-page/custom-page.component.spec.ts
@@ -1,0 +1,27 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { ActivatedRoute } from '@angular/router';
+
+import { CustomPageComponent } from './custom-page.component';
+
+describe('CustomPageComponent', () => {
+  let component: CustomPageComponent;
+  let fixture: ComponentFixture<CustomPageComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [CustomPageComponent],
+      providers: [
+        { provide: ActivatedRoute, useValue: { snapshot: { params: { customPage: '123' } } } }
+      ]
+    })
+      .compileComponents();
+
+    fixture = TestBed.createComponent(CustomPageComponent);
+    component = fixture.componentInstance;
+    //fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});